Greenwood-Nam-D'Agostino (GND) Test on Autopilot

The GND test is based on groupwise error rates for observed versus expected risk, and the test is unstable when any group contains < 5 events (see Demler et al.). The purpose of gnd_test_auto is to automatically check event counts in risk groups and, if necessary, collapse risk groups so that every group has an event count greater than a given threshold.

gnd_test_auto(
  predicted_risk,
  event_time,
  event_status,
  time_predict,
  group_method = "lump",
  group_count_init = 10,
  group_count_min = 2,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 1
)

Arguments

predicted_risk
(numeric vector) predicted risk values for the event of interest at or before a given time.

event_time
(numeric vector) observed event times'

event_status
(numeric vector) observed event status. The values of this vector should be 0 (event censored) and 1 (event observed).

time_predict
(numeric value) the time that risk predictions are computed for. For example, to assess calibration of 10-year predicted risk, set time_predict = 10.

group_method
(character value; 'lump' or 'redo') If 'lump', then a 'lumping' procedure will be applied whenever a group has less than group_min_events_warn events. The lumping procedure will identify whichever group has the lowest event count and assign members of that group to the group with too few events. If 'redo', then the groups will be re-done using cut_percentile but with one less group. 'lump' is the default as this method was studied by Demler et al.

group_count_init
(integer value) the initial number of groups to form based on percentiles of predicted_risk values.

group_count_min
(integer value) the minimum number of groups to attempt running the GND test with. If

group_min_events_warn
(numeric value) The lowest event count within a risk group that will not cause a warning (see details).

group_min_events_stop
(numeric value) The lowest event count within a risk group that will not cause a hard stop (see details).

verbose
(integer value) If 0, no output will be printed. If 1, then output from group reduction will be printed. If 2, then output from group reduction and gnd_test_manual will be printed.

Value

an object of class 'survival.calib_gnd_test'

References

Demler, O.V., Paynter, N.P. and Cook, N.R., 2015. Tests of calibration and goodness‐of‐fit in the survival setting. Statistics in medicine, 34(10), pp.1659-1680. DOI: 10.1002/sim.6428

Examples

# packages
library(riskRegression)

#> riskRegression version 2020.12.08

library(survival)

# set index for training data
train_index <- 1:3000

# fit cox PH model using train data
risk_mdl <- coxph(data = flchain[train_index, ],
                  x = TRUE,
                  formula = Surv(futime, death)~age+sex+flc.grp+lambda)

# set time of prediction
time_predict <- 4000

# predict risk in the testing data
risk_pred <- predictRisk(risk_mdl,
                         newdata = flchain[-train_index, ],
                         times = time_predict)

# run GND test using 10 groups; apply some rules:
#  1. require at least 40 events per group
#  2. try to create 10 groups, and lump the lowest event frequency groups
#     if any groups have less than 40 events.
#  3. Hard stop if this is not obtainable with 5 or more groups.
#  4. set verbose = 2 to see every detail.
# (note that these rules are not a recommendation on how to use the test;
#  this example just shows the mechanics of the automatic group reduction.
#  The recommended values are the default values.)

gnd_test_auto(predicted_risk = risk_pred,
              event_time = flchain$futime[-train_index],
              event_status = flchain$death[-train_index],
              time_predict = time_predict,
              group_min_events_warn = 40,
              group_count_init = 10,
              group_count_min = 5,
              verbose = 2)

#> Checking event counts using 10 risk groups...

#> too few; trying again

#> Checking event counts using 9 risk groups...

#> too few; trying again

#> Checking event counts using 8 risk groups...

#> too few; trying again

#> Checking event counts using 7 risk groups...

#> too few; trying again

#> Checking event counts using 6 risk groups...

#> okay

#> Attempting GND test with 6 risk groups.

#> Group 4 (N = 1950)
#>  - 97 events observed (5.4%)
#>  - 83 events expected (4.3%)
#>  - GND component: 4.383

#> Group 6 (N = 974)
#>  - 82 events observed (9.1%)
#>  - 72 events expected (7.4%)
#>  - GND component: 3.139

#> Group 7 (N = 488)
#>  - 47 events observed (10.5%)
#>  - 49 events expected (9.9%)
#>  - GND component: 0.174

#> Group 8 (N = 487)
#>  - 61 events observed (13.4%)
#>  - 60 events expected (12.3%)
#>  - GND component: 0.457

#> Group 9 (N = 487)
#>  - 68 events observed (15.0%)
#>  - 75 events expected (15.5%)
#>  - GND component: 0.075

#> Group 10 (N = 488)
#>  - 108 events observed (23.6%)
#>  - 113 events expected (23.1%)
#>  - GND component: 0.067

#> -------------------------------------------------------
#> 
#> - Greenwood-Nam-D'Agostino (GND) Mis-calibration Test
#> 
#> -- Chi-square test statistic: 8.295
#> -- degrees of freedom: 5
#> -- P-value for lack of fit: .14
#> -- Warnings: None
#> 
#> -----------------------------------------------------
#>  
#> - Events data by group (truncated):
#> 
#>  group_label group_n events_observed events_expected
#>            4    1950              97            83.1
#>            6     974              82            72.1
#>            7     488              47            48.5
#>            8     487              61            60.0
#>            9     487              68            75.3
#> 
#> + 1 more rows