Greenwood-Nam-D'Agostino (GND) Test

Demler et al developed a modification of the Nam-D'Agostino test that has nominal type 1 error rates with censored outcomes. The GND test is also applicable in settings where the proportional hazards assumption is invalid.

gnd_test_manual(
  predicted_risk,
  event_status,
  event_time,
  time_predict,
  group,
  group_min_events_warn = 5,
  group_min_events_stop = 2,
  verbose = 0
)

Arguments

predicted_risk
(numeric vector) predicted risk values for the event of interest at or before a given time.

event_status
(numeric vector) observed event status. The values of this vector should be 0 (event censored) and 1 (event observed).

event_time
(numeric vector) observed event times'

time_predict
(numeric value) the time that risk predictions are computed for. For example, to assess calibration of 10-year predicted risk, set time_predict = 10.

group
(numeric vector) risk group assignment. The GND chi-square test statistic is based on the difference between predicted and observed event counts in each risk group.

group_min_events_warn
(numeric value) The lowest event count within a risk group that will not cause a warning (see details).

group_min_events_stop
(numeric value) The lowest event count within a risk group that will not cause a hard stop (see details).

verbose
(integer value) If 0, no output will be printed. If 1, details about events in each group will be printed.

Value

an object of class 'survival.calib_gnd_test'

Details

Minimum event counts for risk groups: Low event counts within any risk group may cause high variability in the GND test results. It is recommended that all risk groups have at least 5 events, and this is why the default value of group_min_events_warn is 5. If there are less than 2 events in any group, the GND test is unstable and risk groups should be collapsed. Therefore, the default value of group_min_events_stop is 2.

References

Demler, O.V., Paynter, N.P. and Cook, N.R., 2015. Tests of calibration and goodness‐of‐fit in the survival setting. Statistics in medicine, 34(10), pp.1659-1680. DOI: 10.1002/sim.6428

Examples

library(survival)
library(riskRegression)
train_index <- 1:3000

risk_mdl <- coxph(data = flchain[train_index, ],
                  x = TRUE,
                  formula = Surv(futime, death)~age+sex+flc.grp+lambda)

time_predict <- 4000

risk_pred <- predictRisk(risk_mdl,
                         newdata = flchain[-train_index, ],
                         times = time_predict)

#split into deciles
risk_groups <- cut_percentiles(risk_pred, g = 10)

gnd_test_manual(predicted_risk = risk_pred,
                event_time = flchain$futime[-train_index],
                event_status = flchain$death[-train_index],
                time_predict = time_predict,
                group = risk_groups,
                verbose = 0)

#> -------------------------------------------------------
#> 
#> - Greenwood-Nam-D'Agostino (GND) Mis-calibration Test
#> 
#> -- Chi-square test statistic: 8.84
#> -- degrees of freedom: 9
#> -- P-value for lack of fit: .45
#> -- Warnings: None
#> 
#> -----------------------------------------------------
#>  
#> - Events data by group (truncated):
#> 
#>  group_label group_n events_observed events_expected
#>            1     488              15            14.2
#>            2     487              22            18.8
#>            3     487              25            22.8
#>            4     488              35            27.3
#>            5     487              36            32.7
#> 
#> + 5 more rows